organize files in directories
