+++ src/interface.cpp
@@ -15,7 +15,7 @@
 */
 
 #include <iostream>
-#include "interface.h"
+#include "../include/interface.h"
 
 using namespace std;
